+++ src/rust/src/x509/crl.rs
@@ -580,8 +580,12 @@ fn create_x509_crl(
     private_key: &pyo3::PyAny,
     hash_algorithm: &pyo3::PyAny,
 ) -> CryptographyResult<CertificateRevocationList> {
-    let sigalg = x509::sign::compute_signature_algorithm(py, private_key, hash_algorithm)?;
-
+    let sigalg = x509::sign::compute_signature_algorithm(
+        py,
+        private_key,
+        hash_algorithm,
+        py.None().into_ref(py),
+    )?;
     let mut revoked_certs = vec![];
     for py_revoked_cert in builder
         .getattr(pyo3::intern!(py, "_revoked_certificates"))?
@@ -628,7 +632,13 @@ fn create_x509_crl(
     };
 
     let tbs_bytes = asn1::write_single(&tbs_cert_list)?;
-    let signature = x509::sign::sign_data(py, private_key, hash_algorithm, &tbs_bytes)?;
+    let signature = x509::sign::sign_data(
+        py,
+        private_key,
+        hash_algorithm,
+        py.None().into_ref(py),
+        &tbs_bytes,
+    )?;
     let data = asn1::write_single(&crl::CertificateRevocationList {
         tbs_cert_list,
         signature_algorithm: sigalg,